Extract Eloquent model builder mixin handler
